Waaaaaaaay too many changes for a single commit...
It all started with the new `backlinks` query. This led me to
introducing the DataModel interface, which in turn led me to refreshing
those in parallel.

Since I was introducing concurrency anyway, I also applied it to query
execution.

So, now whenever there's a change detected in the vault, and after they
are applied to the vault, first all datamodels are refreshed in parallel
and when that's done, then all queries are executed in parallel. When
that's done, any documents that have changed are written to disk
sequentially.

Do I need all this concurrency? Probably not. But it's pretty cool.