Made the project list dependent on OmniFocus to sort by priority.
Projects in lists and tables are no longer sorted by last modification
date (according to the timeline), but by priority instead, where the
priority is its relative order in OmniFocus: the higher on the list, the
more important it is.

The OmniFocusRepository has been modified to make sure it runs as fast
as possible, and only blocks threads running through it if a refresh is
happening.