Add support for timeline markers.
A marker is just a normal reference to a document, with superpowers

The `marked` query selects all lines for a document marked with one or
more markers, and shows them together, grouped in sections by marker.
The marker references are removed from the output.

This allows one to move/duplicate important timeline entries out of the
 timeline, into more stateful content.

 See the README for an example.