Add CSV output of trainer data.
The CSV is generated as query output on a page, not to a separate file.

A few more personal data fields are added, to complete the CSV: the
place of residence, and whether a volunteer is also a coach.

With that, the CSV contains all the data necessary for all purposes:

- Generating trainer agreements
- Paying out the trainer compensations
- Checking the trainer certificates of conduct