Replace DataModel refresh ordering mechanism.
Old model: have DataModels return an integer order and use that to
sort all models.

New model: have DataModels return their dependent models and compute
the correct ordering from that.

The new model requires that a DataModel keeps an instance of the
DataModel(s) it depends on. But, that's the logical thing to do, as it
will want to call methods on it. Otherwise why is there a dependency?