Introduce Changelog system, first steps.
The idea is to generalize the handling events by building up a changelog
and have data models read from it, selecting what they find interesting,
and append to it, adding what they did.

In a later step I might even be able to use this to have queries decide
if there was any impact on them, based on the changelog.